+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/pages/Project/Overview/index.js
@@ -9,193 +9,4 @@
  * distribution rights of the Software.
  */
 
-import classNames from 'classnames';
-import {useEffect, useState} from 'react';
-import {useOutletContext} from 'react-router-dom';
-import i18n from '../../../../../common/I18n';
-import {useAppPropertiesContext} from '../../../../../common/contexts/AppPropertiesContext';
-import {getAccountSubscriptions} from '../../../../../common/services/liferay/graphql/queries';
-import CardSubscription from '../../../components/CardSubscription';
-import SubscriptionsFilterByStatus from '../../../components/SubscriptionsFilterByStatus';
-import SubscriptionsNavbar from '../../../components/SubscriptionsNavbar';
-import SupportOverview from '../../../components/SupportOverview';
-import {useCustomerPortal} from '../../../context';
-import {actionTypes} from '../../../context/reducer';
-import {PRODUCT_TYPES, SUBSCRIPTIONS_STATUS} from '../../../utils/constants';
-import {getWebContents} from '../../../utils/getWebContents';
-import OverviewSkeleton from './Skeleton';
-
-const Overview = () => {
-	const [{project, subscriptionGroups}, dispatch] = useCustomerPortal();
-	const {setHasQuickLinksPanel, setHasSideMenu} = useOutletContext();
-	const {client} = useAppPropertiesContext();
-
-	const [accountSubscriptions, setAccountSubscriptions] = useState([]);
-	const [selectedSubscriptionGroup, setSelectedSubscriptionGroup] = useState(
-		''
-	);
-	const [selectedStatus, setSelectedStatus] = useState([
-		SUBSCRIPTIONS_STATUS.active,
-		SUBSCRIPTIONS_STATUS.expired,
-		SUBSCRIPTIONS_STATUS.future,
-	]);
-
-	const [
-		subscriptionGroupsWithSubscriptions,
-		setSubscriptionGroupsWithSubscriptions,
-	] = useState([]);
-
-	const parseAccountSubscriptionGroupERC = (subscriptionName) => {
-		return subscriptionName.toLowerCase().replaceAll(' ', '-');
-	};
-
-	const subscriptionsCards = accountSubscriptions.filter(
-		(subscription) =>
-			subscription.accountSubscriptionGroupERC.replace(
-				`${project?.accountKey}_`,
-				''
-			) === parseAccountSubscriptionGroupERC(selectedSubscriptionGroup) &&
-			selectedStatus.includes(subscription.subscriptionStatus)
-	);
-
-	useEffect(() => {
-		setHasQuickLinksPanel(true);
-		setHasSideMenu(true);
-	}, [setHasSideMenu, setHasQuickLinksPanel]);
-
-	useEffect(() => {
-		const getAllSubscriptions = async (accountKey) => {
-			const {data: dataAccountSubscriptions} = await client.query({
-				query: getAccountSubscriptions,
-				variables: {
-					filter: `accountKey eq '${accountKey}'`,
-				},
-			});
-
-			if (dataAccountSubscriptions) {
-				const dataAllSubscriptions =
-					dataAccountSubscriptions?.c?.accountSubscriptions?.items;
-
-				const accountSubscriptionGroups = subscriptionGroups.filter(
-					(subscriptionGroup) =>
-						dataAllSubscriptions.some(
-							(subscription) =>
-								subscription.accountSubscriptionGroupERC.replace(
-									`${accountKey}_`,
-									''
-								) ===
-								parseAccountSubscriptionGroupERC(
-									subscriptionGroup.name
-								)
-						)
-				);
-
-				setAccountSubscriptions(dataAllSubscriptions);
-
-				setSubscriptionGroupsWithSubscriptions(
-					accountSubscriptionGroups.sort(
-						(
-							previousAccountSubscriptionGroup,
-							nextAccountSubscriptionGroup
-						) =>
-							previousAccountSubscriptionGroup?.tabOrder -
-							nextAccountSubscriptionGroup?.tabOrder
-					)
-				);
-			}
-		};
-
-		if (subscriptionGroups && project) {
-			getAllSubscriptions(project.accountKey);
-		}
-	}, [client, project, subscriptionGroups]);
-
-	useEffect(() => {
-		if (project && subscriptionGroups) {
-			dispatch({
-				payload: getWebContents(
-					project.dxpVersion,
-					project.slaCurrent,
-					subscriptionGroups
-				),
-				type: actionTypes.UPDATE_QUICK_LINKS,
-			});
-		}
-	}, [dispatch, project, subscriptionGroups]);
-
-	if (!project || !subscriptionGroups) {
-		return <OverviewSkeleton />;
-	}
-
-	const isPartnership =
-		selectedSubscriptionGroup === PRODUCT_TYPES.partnership ||
-		subscriptionGroups[0]?.name === PRODUCT_TYPES.partnership;
-
-	return (
-		<>
-			<SupportOverview koroneikiAccount={project} />
-			<div className="d-flex flex-column mr-4 mt-6">
-				{!isPartnership && <h3>{i18n.translate('subscriptions')}</h3>}
-
-				{!!subscriptionGroupsWithSubscriptions.length && (
-					<>
-						<div
-							className={classNames('align-items-center d-flex', {
-								'justify-content-between':
-									subscriptionGroupsWithSubscriptions.length <
-									5,
-								'justify-content-evenly':
-									subscriptionGroupsWithSubscriptions.length >
-									4,
-							})}
-						>
-							<SubscriptionsNavbar
-								selectedSubscriptionGroup={
-									selectedSubscriptionGroup
-								}
-								setSelectedSubscriptionGroup={
-									setSelectedSubscriptionGroup
-								}
-								subscriptionGroups={
-									subscriptionGroupsWithSubscriptions
-								}
-							/>
-
-							<SubscriptionsFilterByStatus
-								selectedStatus={selectedStatus}
-								setSelectedStatus={setSelectedStatus}
-							/>
-						</div>
-
-						<div className="cp-overview-cards-subscription d-flex flex-wrap mt-4">
-							{subscriptionsCards.length ? (
-								subscriptionsCards.map(
-									(accountSubscription, index) => (
-										<CardSubscription
-											cardSubscriptionData={
-												accountSubscription
-											}
-											key={index}
-											selectedSubscriptionGroup={
-												selectedSubscriptionGroup
-											}
-										/>
-									)
-								)
-							) : (
-								<p className="mx-auto pt-5">
-									{i18n.translate(
-										'no-subscriptions-match-these-criteria'
-									)}
-								</p>
-							)}
-						</div>
-					</>
-				)}
-			</div>
-		</>
-	);
-};
-
-Overview.Skeleton = OverviewSkeleton;
-export default Overview;
+export {default} from './Overview';
